2022/11-monkey-in-the-middle/main_test.go
package main

import (
	"fmt"
	"testing"
)

var example = `Monkey 0:
  Starting items: 79, 98
  Operation: new = old * 19
  Test: divisible by 23
    If true: throw to monkey 2
    If false: throw to monkey 3

Monkey 1:
  Starting items: 54, 65, 75, 74
  Operation: new = old + 6
  Test: divisible by 19
    If true: throw to monkey 2
    If false: throw to monkey 0

Monkey 2:
  Starting items: 79, 60, 97
  Operation: new = old * old
  Test: divisible by 13
    If true: throw to monkey 1
    If false: throw to monkey 3

Monkey 3:
  Starting items: 74
  Operation: new = old + 3
  Test: divisible by 17
    If true: throw to monkey 0
    If false: throw to monkey 1`

func Test_part1(t *testing.T) {
	tests := []struct {
		name  string
		input string
		want  int
	}{
		{
			name:  "example 1",
			input: example,
			want:  10605,
		},
		{
			name:  "actual",
			input: input,
			want:  55944,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part1(tt.input); got != tt.want {
				t.Errorf("part1() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_part2(t *testing.T) {
	tests := []struct {
		name  string
		input string
		want  int
	}{
		{
			name:  "example 1",
			input: example,
			want:  2713310158,
		},
		{
			name:  "actual",
			input: input,
			want:  15117269860,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part2(tt.input); got != tt.want {
				t.Errorf("part2() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_performOperation(t *testing.T) {
	tests := []struct {
		item         int
		rawOperation string
		want         int
	}{
		{3, "+ 5", 8},
		{3, "* 5", 15},
		{3, "* old", 9},
	}
	for _, tt := range tests {
		name := fmt.Sprintf("%v %v want %v", tt.item, tt.rawOperation, tt.want)
		t.Run(name, func(t *testing.T) {
			if got := performOperation(tt.item, tt.rawOperation); got != tt.want {
				t.Errorf("part2() = %v, want %v", got, tt.want)
			}
		})
	}
}
